adjust id precison for shader profiling

+++ CubeEngine/CubeEngine/Render/Shader/DBModel/CEShaderStructInfo.h
@@ -13,7 +13,7 @@
 
 @interface CEShaderStructInfo : NSObject <CEJsonCoding>
 
-@property (nonatomic, readonly) unsigned long long structID;
+@property (nonatomic, readonly) uint32_t structID;
 @property (nonatomic, readonly) NSString *name;
 @property (nonatomic, readonly) NSArray *variables;
