+++ CubeEngine/CubeEngine/Render/Shader/DBModel/CEShaderVariableInfo_setter.h
@@ -10,7 +10,7 @@
 
 @interface CEShaderVariableInfo ()
 
-@property (nonatomic, assign, readwrite) unsigned long long variableID;
+@property (nonatomic, assign, readwrite) uint32_t variableID;
 @property (nonatomic, strong, readwrite) NSString *name;
 @property (nonatomic, assign, readwrite) NSString *type;
 @property (nonatomic, strong, readwrite) NSString *precision;
